refactor: change instance to session.ts
